Feature/sc 353/soportar logs no evtx (#62)
* partial commit

* Update log.js

* Update log.js

* Update index.js

* Merge branch 'develop' into feature/sc-348/crud-de-timeline-edicion-de-eventos-backend

* commit parcial

* fixes from qa

* fixes from pr

* Update log.js

* Update line.js

+++ evtx-converter/converter/index.js
@@ -56,8 +56,9 @@ const convertFile = async (req, resp) => {
         } else {
           throw { code: 400, msg: 'Only linux platform is supported' }
         }
+        return { input: inputFileName, output: convertedName, metadata }
       }
-      return { input: inputFileName, output: convertedName, metadata }
+      return { input: inputFileName, metadata }
     }))
     // Send to server
     const url = `${process.env.HOST_CORRELATION}${req.path}`
@@ -81,7 +82,9 @@ const convertFile = async (req, resp) => {
     }).finally(() => {
       files2send.forEach(({ input, output }) => {
         fs.unlinkSync(input)
-        fs.unlinkSync(output)
+        if(output) {
+          fs.unlinkSync(output)
+        }
       })
     })
